pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "protovalidate"
description = "Protocol Buffer Validation for Python"
readme = "README.md"
license = "Apache-2.0"
license-files = ["LICENSE"]
keywords = ["validate", "protobuf", "protocol buffer"]
requires-python = ">=3.9"
classifiers = [
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Operating System :: OS Independent",
  "Typing :: Typed",
]
dynamic = ["version"]
dependencies = [
    "protobuf==6.*",
    "cel-python==0.2.*",
]

[project.urls]
Homepage = "https://github.com/bufbuild/protovalidate-python"
Source = "https://github.com/bufbuild/protovalidate-python"
Issues = "https://github.com/bufbuild/protovalidate-python/issues"

[dependency-groups]
dev = [
	"mypy",
	"ruff",
	"types-protobuf==6.30.2.20250503",
]

[tool.hatch.version]
source = "vcs"

[tool.ruff]
line-length = 120
lint.select = [
	"A",
	"ARG",
	"B",
	"C",
	"DTZ",
	"E",
	"EM",
	"F",
	"FBT",
	"I",
	"ICN",
	"N",
	"PLC",
	"PLE",
	"PLR",
	"PLW",
	"Q",
	"RUF",
	"S",
	"T",
	"TID",
	"UP",
	"W",
	"YTT",
]
lint.ignore = [
	# Allow boolean positional values in function calls, like `dict.get(..., True)`.
	"FBT003",
	# Ignore complexity
	"C901",
	"PLR0911",
	"PLR0912",
	"PLR0913",
	"PLR0915",
	# Ignore magic values - in this library, most are obvious in context.
	"PLR2004",
]
lint.unfixable = [
	# Don't autofix unused imports.
	"F401",
]

[tool.ruff.lint.isort]
known-first-party = ["protovalidate", "buf"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.per-file-ignores]
# Tests can use magic values, assertions, and relative imports.
"tests/**/*" = ["PLR2004", "S101", "TID252"]
